+++ src/ua/tasks/IntermediateWebProject/src/main/java/ua/training/model/bean/RealEstateInsurance.java
@@ -6,7 +6,7 @@ public class RealEstateInsurance extends LiabilityInsurance {
     private String name;
     private BigDecimal valuationOfPropertyPrice;
 
-    public RealEstateInsurance(Long id, BigDecimal liabilityPrice,
+    public RealEstateInsurance(Integer id, BigDecimal liabilityPrice,
                                Double risk, String name, BigDecimal valuationOfPropertyPrice) {
         super(id, liabilityPrice, risk);
         this.name = name;
